# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.WorkspaceStatus:
    enum:
    - active
    - deleted
    type: string
    x-enum-varnames:
    - WorkspaceStatusActive
    - WorkspaceStatusDeleted
...
  handler.createWorkspaceRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  handler.updateWorkspaceRequest:
    properties:
      description:
        type: string
      name:
        type: string
      status:
        $ref: '#/definitions/domain.WorkspaceStatus'
    type: object
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 科研项目文档管理系统后端 API。
  title: Research Admin Backend API
  version: "1.0"
paths:
...
  /workspaces:
    get:
      description: 返回当前登录用户加入的 workspace 列表。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取 Workspace 列表
      tags:
      - workspaces
    post:
      consumes:
      - application/json
      description: 创建 workspace，并把当前登录用户设置为 owner。
      parameters:
      - description: Workspace 参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createWorkspaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建 Workspace
      tags:
      - workspaces
  /workspaces/{workspaceId}:
    delete:
      description: 软删除 workspace，仅 workspace owner 可操作。
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 删除 Workspace
      tags:
      - workspaces
    get:
      description: 返回 workspace 根目录或指定父文档的一层直接子节点，并按当前用户权限过滤。
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: 父文档 ID
        in: query
        name: parentId
        type: string
      - description: 文档状态
        in: query
        name: status
        type: string
      - description: 返回数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取 Workspace 目录子节点
      tags:
      - workspaces
    patch:
      consumes:
      - application/json
      description: 更新 workspace 元数据，仅 workspace owner 可操作。
      parameters:
      - description: Workspace ID
        in: path
        name: workspaceId
        required: true
        type: string
      - description: Workspace 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateWorkspaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新 Workspace
      tags:
      - workspaces
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"